Update recursion.php
recursion vs. iterative:
---------------------------note-taking-----------------------------
Recursion is a method that is executed continuously like a loop until the condition is met. Recursions can be replaced by using loops. The main difference that I know is when using loops, you must specify the number of times that method is being executed and with recursion, you must have at least one base case to stop executing, or else it will not end which can lead to errors or crashing the program.
One advantage of using recursion is that writing up recursion methods/functions are a lot more simplex for a basic argument than taking up an iterative approach. Recursions only need one condition to break the loop. It could also save lines of code to achieve the same output as a loop could.
Though recursion can be easily replaced by an iterative / loop approach, I think the best way to use recursion is to help with skilling up your problem-solving skills. Recursion breaks down the problem, allowing you to understand how stacks work such as (last in, first out) and making use of them.